Adding the DirectReports GET component
