+++ samples/react-contoso/.env.sample
@@ -3,7 +3,4 @@ SKIP_PREFLIGHT_CHECK=true
 FAST_REFRESH = false
 REACT_APP_SITE_NAME="Contoso Dashboard"
 REACT_APP_CLIENT_ID="ed072e38-e76e-45ae-ab76-073cb95495bb"
-REACT_APP_INCIDENTS_LIST_ID="45d48008-6b4f-4580-9a05-b23b21afa578"
-REACT_APP_USER_INFORMATION_LIST_ID="3c781595-6e17-404b-a6fc-fcb76c3b9164"
-REACT_APP_SUPPORT_GROUP_ID="a3bf7b91-b1da-42de-bb00-78b35dee96db"
